REFACTOR: codes structure primitive (#558)
* FEAT: EDB configuration create new stackup

* FIX

* MISC: Auto fixes from pre-commit.com hooks

For more information, see https://pre-commit.ci

* FIX

* Update src/pyedb/dotnet/edb_core/cell/primitive.py


* Update src/pyedb/dotnet/edb_core/cell/primitive.py


* Update src/pyedb/dotnet/edb_core/cell/primitive.py


* Update src/pyedb/dotnet/edb_core/cell/primitive.py


* Update src/pyedb/dotnet/edb_core/cell/primitive.py


* MISC: Auto fixes from pre-commit.com hooks

For more information, see https://pre-commit.ci

* Update src/pyedb/dotnet/edb_core/cell/primitive.py

---------
