+++ src/pyedb/dotnet/edb_core/definition/component_def.py
@@ -23,7 +23,7 @@
 import os
 
 from pyedb.dotnet.edb_core.definition.component_model import NPortComponentModel
-from pyedb.dotnet.edb_core.obj_base import ObjBase
+from pyedb.dotnet.edb_core.utilities.obj_base import ObjBase
 from pyedb.generic.general_methods import pyedb_function_handler
