+++ src/pyedb/dotnet/edb_core/definition/package_def.py
@@ -21,7 +21,7 @@
 # SOFTWARE.
 
 from pyedb.dotnet.edb_core.geometry.polygon_data import PolygonData
-from pyedb.dotnet.edb_core.obj_base import ObjBase
+from pyedb.dotnet.edb_core.utilities.obj_base import ObjBase
 from pyedb.generic.general_methods import pyedb_function_handler
