+++ src/pyedb/dotnet/edb_core/utilities/obj_base.py
